% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpca_cache.R
\name{.tpca_cache_entry_dir}
\alias{.tpca_cache_entry_dir}
\title{Return the cache directory for a specific hash key}
\usage{
.tpca_cache_entry_dir(key)
}
\description{
Return the cache directory for a specific hash key
}
\keyword{internal}
