% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpca_cache.R
\name{.tpca_cache_exists}
\alias{.tpca_cache_exists}
\title{Check whether a cache entry exists and is complete}
\usage{
.tpca_cache_exists(key)
}
\description{
Check whether a cache entry exists and is complete
}
\keyword{internal}
