% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpca_bridge.R
\name{get_tpca_results}
\alias{get_tpca_results}
\title{Retrieve TPCA fit metadata from a Seurat object}
\usage{
get_tpca_results(obj, contour_type = NULL)
}
\arguments{
\item{obj}{A Seurat v5 object.}

\item{contour_type}{\code{"cell"} or \code{"nucleus"}. \code{NULL} returns
the full tpca misc list.}
}
\value{
A list of TPCA fit metadata.
}
\description{
Retrieve TPCA fit metadata from a Seurat object
}
\seealso{
\code{\link{store_tpca_results}}
}
