% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpca_bridge.R
\name{store_tpca_results}
\alias{store_tpca_results}
\title{Store TPCA results in a Seurat object}
\usage{
store_tpca_results(obj, tpca_result, reduction_key_prefix = "ShapePC_")
}
\arguments{
\item{obj}{A Seurat v5 object.}

\item{tpca_result}{The list returned by \code{\link{run_tpca}}.}

\item{reduction_key_prefix}{Key prefix for the \code{DimReduc}. Default
\code{"ShapePC_"}.}
}
\value{
The Seurat object with TPCA results stored.
}
\description{
Writes the PGA embedding as a \code{DimReduc} slot and fit metadata to
\code{obj@misc$kstitch$tpca[[contour_type]]}.
}
\seealso{
\code{\link{run_tpca}}, \code{\link{get_tpca_results}}
}
